Make Windows build use $PSScriptRoot
This converts the build script for Event Store on Windows to run from any
invocation path rather than relying on it being executed from the root of
the directory structure by using $PSScriptRoot in common with the
deployment scripts.

NOTE: This introduces a dependency on PowerShell 3 minimum, but it's quite
likely that people building from source will already have this.

In addition, now we have a top level scripts directory it makes more sense
for the powershell build script to be in scripts/ rather than
/tools/powershell.