Allow the recreation of deleted projections (same name)
When deleting a projection, ensure that a $ProjectionDeleted event is
written to $projections-$all
This event allows us to know which projections are actually alive.

When creating a projection, we check for the existence of a
$projections-[projection-name] stream. If this stream exists, we grab the
last known event number, increment it by one and that would be the version
for the projection.

Checkpoint stream for projection should be deleted when the user
requests for it to.

Should not hard delete the checkpoint stream (there might be a case where
the user wants to recreate the projection)

When loading the checkpoint, there might be a case where there are no
events as the stream has been deleted.

However, if we attempt to write to that checkpoint stream and tell it that we expect
the stream to not exist (Expected Version: -1), it will fail. We need to
cater for situations where the checkpoint stream is being reused for
instances where the projection gets recreated.