Change default log and data path based on platform
As part of the package builds for Linux it's important we use the correct
default paths for the OS distributions we target. The previous defaults
were to create a new temporary database and log directory if none were
specified.

This commit changes that to use a subdirectory of the execution path on
Windows and Mac OS X, and to use /var/lib/eventstore and
/var/log/eventstore for data and logs respectively on Linux.

It also introduces the constants needed for putting static content in the
correct places on the file system for Linux while keeping the Windows and
Mac OS X distributions the same.