Key was not present in dictionary This issue comes up when the node has been issued with a shutdown command and the following chain of events occur.
On shutdown, the ProjectionCoreCoordinator receives the a state change
message. Now given that it was started, it would be requested to stop.

This stop would then in turn publish a StopCore as well as a StopReader
(If run projections System or All)

The StopCore message will be received by the ProjectionCoreService and
will in turn attempt to kill each of the Projections. Each of these
projections would be asked to asked to ensure that they have been
unsubscribed (CoreProjection::EnsureUnsubscribed) which then publishes
an ReaderSubscriptionManagement.Unsubscribe message.

There are 2 messages to take note of here. StopCore and StopReader.
StopCore may result in an Unsubscribe message for each of the
projections.

StopReader was published from the ProjectionCoreCoordinator. StopReaders will essentially clear out the lists such as subscriptions.

Then along comes the Unsubscribe message for the projection and since
the StopReaders cleared out the list, the subscriptionId will not
longer be in the list, as it's empty.