Add script for packaging from Xamarin Mono package
This commit adds a new script, package-mono-xamarin.sh, which allows
creation of a statically linked Event Store package with Mono installed
from the Xamarin Debian repository.

Unfortunately there are some assembly config files which are present in
the packages which are not present if you build from source, which then
cause what appears to be a bug in mkbundle to be hit whereby it
generates C code which declares arrays of unsigned char which are then
passed into a function expecting const char *. This produces warnings
under gcc 4.8.

Rather than fix mkbundle (which would take time to be reflected in the
package repositories, if it were even accepted at all), we instead pass
the generated source through sed, removing the unsigned modifier for the
path to the configuration files. The source then compiles without
warnings.