Move -i"" to beginning for BSD sed compatibility
MacOS has BSD sed which doesn't like having the -i"" after the edit
command as it takes - to be the end of the input and therefore reads
from standard in, which is an error with -i (at least I think that's
what's happening). Problem does not occur if the -i"" is before the edit
command and filename.

This has not been an issue as our build boxes have GNU sed installed via
homebrew with `--use-default-names`, as all Macs really should...