Publishes scavenge status events from StorageScavenger and TFChunkScavenger
Creates three new events: ScavengeStarted, ScavengeChunksCompleted and ScavengeCompleted.

Between the StorageScavenger and TFChunkScavenger, two streams will be produced.
The first is a details stream which is created per scavenge instance.
The second is an index stream which has linked events to the ScavengeStarted and ScavengeCompleted events from the instance stream.

Adds ScavengeHistoryMaxAge command line argument to set the number of days to keep scavenge history.