Handles Access Denied exceptions when setting up the DB with the default data directory
Handles Unauthorized Access Exceptions caught while setting up the DB.
If the db path is the default data directory, it changes the dbPath to a fallback directory.
If the db path is something else, then the error is thrown.

Adds a new Location called FallbackDefaultDataDirectory which is set to ./data