Replace DebugLogger with Common Event Store Logging Ensure that the user management service only notifies once that it's been initialized Introduce a SystemReady message
On user management service initialized, start an internal init.
Once the subsystems have been started, fire off a SystemReady message.

Ensure that the subsystems are only started once on initialization

Cancel the async operations when needed

When a node comes back up as Master, it re-subscribes and since the
cancellation token is never invoked, duplicate messages gets received
which puts projections into bad states

Check if the subscription exists before getting it's value

Adding the event numbers helps with debugging issues, for instance finding duplicate messages

Add relevant logging to key points in projections

On the starting of services, reset the writers.

The writers seem to get themselves stuck and in a bad state when the nodes
shift statuses. Reset them as part of the node status shift.

Dont attempt to kill projections, but rather just clean them up in memory
such as removing them from the list.
TODO: Check for proper cleanup

Projection Tests Subscribe and publish SystemMessage.SystemReady message.

Handle reading from soft deleted streams
When soft deleting streams, the reader will attempt to read from 0 and
expect the first event number to be zero, but this is not the case in soft
deleted streams.

Remove unused Logger and correct spelling