Only initialise the existing projections when ready
There are cases when the core worker readers takes a while to start up and
by that time the manager could have started the projections and messages
could have been written to the core worker's stream and they could have
been missed. The reason they can be missed is that when the core reader
starts, it will read the last event in the $projections-${worker_id} stream
and marks that as the starting point to read from.