Track and deleting of emitted streams The emitted streams can be tracked if the user indicates that he/she wants to do so, this is an option when creating a projection. When the user wants to delete a projection he/she can indicate that they want to delete all of the emitted streams that resulted from the projection.
The ManagedProjection has also been reviewed and cleaned up as part of
this exercise

Track and deleting of emitted streams
The emitted streams can be tracked if the user indicates that he/she wants
to do so, this is an option when creating a projection.
When the user wants to delete a projection he/she can indicate that they
want to delete all of the emitted streams that resulted from the
projection.

The ManagedProjection has also been reviewed and cleaned up as part of
this exercise