Fix merging of 32 to 64 bit if hash collisions
There is an issue that occurs during a merge operation when a 32bit index
is merged with a 64bit index and the 32bit index contains hash collisions.

This commit moves the logic that was spread across both types of Merge
methods into the EnumerableTable.

The EnumerableTable will for a 32bit index now read chunks of entries,
upgrade them and place them in a sorted memory index that is used in the
merge.

The chunks are defined by reading the PTable until a different hash is
encountered.