Fix 2 additional duplicate entry issues
When the duplicate entry is in the memory table.
---

When reading backwards or forwards in the index reader there is a
case where the duplicate index entry could be in the mem table (which is
64bit by default) and the rest of the index entries are 32bit. This
means that instead of the range query on the index returning something
along the lines of

Stream Hash  Version    Position
111          2          3
111          1          2
111111       0          4 (duplicate)
111          0          1

it will return something like the following

Stream    Version    Position
111111    0          4 (duplicate)
111       2          3
111       1          2
111       0          1

Proposed Solution:
The IndexReader::ReadStreamEventsForward and
IndexReader::ReadStreamEventsBackward  will now order by version
for the entries returned from the index so that the correct
non-duplicate entry are selected.

When rebuilding the memory table on node startup and the duplicate entry
is cached as the last event number.
---

When a node starts up, it will rebuild the memory index and cache the
event numbers as it reads. This can cause an issue in the case
where the duplicate event exists which will result in the incorrect last event
number being cached for the stream.

Proposed Solution:
- Disable the caching when building up the memory index.
- The additional checks are also disabled during index rebuilding.
    Note: as of this time, the additional checks aren't running in
    production, only for tests.
- IndexReader::GetStreamLastEventNumberUncached will now always look one
ahead and determine if it's really the latest event number for the
stream.

Note:
There is a possibility that the above mentioned looking one ahead will
not find the "real" latest event number, as the user could have written many
duplicate events to the stream.

Running a scavenge should get rid of all of the duplicates.