Fault the projection on failed subscription handle
There are cases where the projection subscription has failed to handle a
distributed event and in those cases, the other projection subscriptions
would not have received the event and in those cases, those projection
subscriptions would have "skipped" those events.

The solution here is to fault the projection and cancel it's
subscription.

The other part of this is that the projection would need it's checkpoint
amended so it can continue processing or we can include an option
whereby we let the projection subscriptions skips these events.