Amend the prepare position returned from a prepare
There is a case where a read could happen before a transaction's commit
record has been written and where a single write could appear after the
transaction end and in this case, given the current next position that
is returned from the reader, the next read from that position would skip
the transaction's prepares as they have lower positions than that
returned from the previous read.

It is safe to use 0 as the prepare position for a position returned from
a committed prepare in this case as we use the provided commit position
to move the reader to where we will be reading from an upon reading a
commit, we will use the transaction position to adjust the reader
position, however we still use the provided position to validate
whether the prepare is one we actually care about.