Cleanup for TCP Client Dispatcher
Remove V1ClientDispatcher and revert TcpService, TcpCommand and ConnectionManager to previous versions.
Add versioned registry of wrappers/unwrappers and pass version in when wrapping/unwrapping messages.
Add wrappers and unwrappers for tcp messages that need to be downgraded for version 1 clients.
Remove unnecessary proto messages.

Send an identify message when a client tries to connect that contains the client's version and connection name.
If identify on client fails, the client will try to reconnect.