Change event numbers from integer to long.
These were missed from the original 64bit event number upgrade

- Although the StreamWriter is only used for a specific set of tests that
doesn't care about the 64bit event versions, it needs to be changed to
be consistent with the rest of the codebase.
- The TestFixtureWithExistingEvents's GetTfPos and Assert Event also falls
under the same category as above where it's used for a narrow spectrum
of tests but should be changed to be consistent with the rest.
- The PersistentSubscriptionCheckpointReader was still attempting to see
it's checkpoints as ints instead of longs. There is a ticket for writing
tests around this.
- Remove the IndexCache is it's not being used.
- The Atom Controller still expected integer event numbers in certain
cases.
- Http Convert was expecting a linkTo's event number to be an integer
still.