Rework the the emitted stream to batch up writes
Originally the emitted stream would attempt to write as soon as it
received an emitted event(s). In a scenario where a projection emits to
many different streams, it would flood event store with writes for as
many streams as it was emitting to.

This would cause a series of commit
timeouts to occur and eventually it will fault the projection due to the
maximum number of retries for a write being hit.