#1363 Fix ReadEventsInternalAsync and ReadEventsCallbackAsync called each other recursively. This resulted in stackoverflowexception when subscribing to a very large stream and a lot of memory pressure.
Main fix is to create a do..while in ReadEventsInternalAsync to facilitate continuing with the next slice if not done. This in stead of the recursive call.
To make this work ReadEventsCallbackAsync must return a bool indicating if we are done (shouldstop is true OR ProcessEventsAsync indicates that we are done)

As a side effect a couple of parameters could be dropped because they were only passed between the functions but not really used.