Optimize existAt calls on scavenged chunks using bloom filters for faster index merges.
Description
-----------
For each scavenged chunk, we now keep an associated bloom filter that's used to keep all the log positions from the chunk's PosMap.
When --optimize-index-merge is set to true, ExistAt calls on the chunk look up the log position in the bloom filter (in memory) instead of the previous method that did a binary search lookup on the cached PosMap midpoints (which also requires disk reads)
ExistAt may now return false positives with a low probability (currently set to 0.0001) but it does not affect the operation of index merges.
False positives will only cause a few index entries for events that have been deleted to stay in the PTable but they will be verified against the chunk file in any case.

To limit the amount of memory used by the bloom filters, TFChunkReaderExistsAtOptimizer manages the number of chunks with cached bloom filters to keep in memory (currently set to 10000 chunks, around 5GB max)
TFChunkReaderExistsAtOptimizer uses an LRU cache to remove the least-recently accessed chunk (based on ExistAt calls) when the cache is full.

Detailed list of changes
------------------------
Added an option --optimize-index-merge that triggers the optimization
Added a bloom filter implementation
Added a new class TFChunkReaderExistsAtOptimizer that bounds number of bloom filters to keep in memory
Optimize Posmap reads by reducing the number of seeks
Added _onPut, _onRemove callback to LRUCache
Added tests for TFChunkReaderExistsAtOptimizer, BloomFilter