+ [Server Side Filtering] Now counting filtered out events as well
Problem: When filtering out events and trying to assemble a whole
batch of events, the operation is unbound timewise. It is possible that
we might need to read the whole all stream and still can't get a whole
batch of events together. In this case, clients will timeout.

Current solution: The client decides, if the stream is at an end, if
the server sends an empty list of events. So we can't just return
an empty events list. So, if we have read a batch size of events and
don't have any events to return to the client, we just return the last
event we read, regardless of filter. This ensures the client will
continue to read. It also breaks the filtering. But if we consider
the filtering to be best effort, we might get away with it. Considering
we are free to define the filter option in the way we need.