+ [Server Side filtering] Removed custom code from ReadAllEventsForward and added new command ReadAllEventsForwardFiltered
This enables some behaviour changes:
- Returning an empty list of events is now legal and no longer is
interpreted as end-of-stream
- Instead of best-guessing a search window, the client can give
us this. This enables clients to find their best trade-off between
command times and reading as fast as possible.

Lots of duplication in this commit. It is not straight forward to do
code sharing, as most new methods have new parameter types
or return types. Also, the client messages can't be subclassed,
as this leads to the wrong handler being called for subclasses.