Depose Master if the priority has changed
Deposing the master will kick off elections. This has a number of
benefits instead of kicking off elections manually.

- You won't potentially end up in a case where too many master's are
alive.
- Client's that prefer master will automatically reconnect to the new
master node as we handle this case in the state machine
(ClusterVNodeController) as the node state is now `Unknown`.
- The other nodes in the cluster will realize via their tcp connections
being dropped that the cluster has changed and will act accordingly.