Expose resignation of Master over HTTP. (Previously known as Maintenance Mode) (#1982)
Add ability to issue a resignation of the master node over http.

Allow setting the priority for a node during runtime.
Add NodePriority to ElectionMessage.Proposal.

Add new ResigningMaster state. The node will enter this state when it is told to resign.
While resigning, the master will ignore any new write requests and wait until the request queue is drained.
Once the request queue is empty, the master will enter the Unknown state.

Broadcast the resigning master message to other nodes in the cluster.
This provides nodes with a hint to not re-elect the last master if it has signaled that it is resigning.