Introduce election behaviour tests (#2038)
Improve tests around elections

- Rename NonLeader to Acceptor.

- Exclude ReadOnlyReplica from acknowledging a master resignation.
It does not vote in elections.

- Exclude ReadOnlyReplica from the list of servers to which a
leader for the current round of elections is picked. Since read only
replicas dont handle prepare oks which is sent as replies for the
prepares.

- Rename TimeProvider Now to UtcNow and introduce LocalTime to TimeProvider. 
Instead of introducing another component that tracks utc now, we can
leverage the existing time provider.