Ignore the $type property when deserializing
Customers are using $type as a means to encode some information in their
messages persisted in Event Store. They don't and should not care about
what we use internally and therefor we should do as little as possible
to mess with the data they provide us.

Setting the MetadataPropertyHandling to Ignore will tell Newtonsoft Json
to ignore these metadata properties when deserializing.