Drop macOS as a supported server OS in CI
Since we have switched client access to gRPC, we have been unable to
support macOS as a server target without disabling TLS, thanks to a lack
of support for server-side Application-Layer Protocol Negotiation (ALPN)
in Security.framework, and a lack of willingness on our part to commit
to doing something like binding to OpenSSL on macOS.

This is tracked in grpc/grpc-dotnet#416.

Consequently we've decided to remove the build for macOS for the Event
Store server, which this commit does. Note the _client_ will still need
to be supported and tested on macOS, and we will do this by moving the
gRPC client into a submodule of the server.