cleanly release ExclusiveDbLock
As IHostedService StartAsync / StopAsync are not guaranteed to run on the same thread you can see 'Object synchronization method was called from an unsynchronized block of code.' Further manual testing (on Linux) reveals that this is not a problem in practice. However, this needs to be tested manually on windows and OSX.

e.g. dotnet run --dev --db=$dbpath --mem-db=false

Then kill with both ctl-c and kill -9, then start EventStore again.