Generalize TF and Index in preparation for LogV3
In LogV2 the event records (prepare records) contain the stream name as a string. In LogV3 the stream will be identified by a long.
LogV3 also has a different record header format, which does not include the logical position of each record.
We'd prefer to keep using the existing Index and TF machinery for now since it is mature.

This PR generalises that machinery to support both log formats and their different ways of identifying streams.

I've tried to make the changes as obvious as possible, avoiding refactoring or whitespace adjustments in order to make the changes
as easy to read as possible, since it does affect quite a few lines.