Update standard backup procedure
- Added support for including the bloom filter in the backup
- Fixed bug where checkpoints wouldn't be updated on windows (-I flag: see below)
- Fixed bug where index directory would be copied into index directory resulting in backup/index/index/

The -I flag is required on windows for memory mapped files, because the modified date is not reliably updated. We could add a call to set the modified date excplicitly when flushing and remove this flag.
Note that rsync will still be fairly efficient because of its delta transfer algorithm.