Don't read epochs if they should be cached
- Prevent the EpochManager from attempting to read epochs that
should have been cached.
- Handle read errors in EpochManager.IsCorrectEpochAt
- Add tests for EpochManager and LeaderReplicationService

Since epochs are numbered according to the election proposal that
created them, it is possible for there to be gaps in epoch numbers.

The EpochManager would treat these gaps as cache misses and attempt to
read the epoch. This resulted in an invalid read which would prevent a
replica from subscribing.